% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_processing.R
\name{check_chunk_codefig}
\alias{check_chunk_codefig}
\title{Check for codefig chunks}
\usage{
check_chunk_codefig(chunk_name = chunk_name)
}
\arguments{
\item{chunk_name}{Chunk options vector}
}
\value{
Logical vector of chunks that can be safely unnamed
}
\description{
Check to see which chunks do not have codefig in their label
}
\details{
We want to make sure chunks relying on \code{fig_chunk()} remain named,
so we don't re-name them and then break the label that allows left-code / right-plot slides to be constructed.
By design every \code{fig_chunk()} dependent chunk should have \code{codefig} in its label. (This is managed via snippets.)
}
