% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_processing.R
\name{check_chunk_include}
\alias{check_chunk_include}
\title{Which chunks are safe to exclude}
\usage{
check_chunk_include(chunk_options = chunk_options)
}
\arguments{
\item{chunk_options}{Chunk options vector}
}
\value{
Logical vector of chunks that can be safely unnamed
}
\description{
Check to see which chunks do not have include = FALSE
}
\details{
We want to make sure chunks invovled in flipbookr remain named,
so we don't re-name them and then break the label that allows flipbooks to be constructed.
}
