% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_theme.R
\name{get_ggplot_colors}
\alias{get_ggplot_colors}
\title{Generate ggplot default colors}
\usage{
get_ggplot_colors(n = 6, h = c(0, 360) + 15)
}
\arguments{
\item{n}{Number of colors}

\item{h}{Hue}
}
\value{
n colors every 15 degrees around the color wheel
}
\description{
Generate ggplot default colors
}
