% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{kjh_clear_registry}
\alias{kjh_clear_registry}
\title{Clear the systemfonts registry}
\usage{
kjh_clear_registry()
}
\value{
Setting
}
\description{
Clear the systemfonts registry
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
