% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templating.R
\name{kjh_css_files}
\alias{kjh_css_files}
\title{Construct vector of CSS file paths}
\usage{
kjh_css_files(
  css_files = c("kjh-slides.css", "tenso-berkeley.css", "animate.css", "widths.css"),
  css_dir = "css"
)
}
\arguments{
\item{file}{Vector of custom CSS files}

\item{path}{Where the files are in the package}
}
\value{
Vector of css files to use for the css: arg in moon_reader
}
\description{
Construct vector of CSS file paths
}
