% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_slides.R
\name{kjh_decktape_all}
\alias{kjh_decktape_all}
\title{Render every HTML file in the \verb{slides/} folder to PDF and put it in \verb{pdf_slides/}}
\usage{
kjh_decktape_all(indir = "slides", outdir = "pdf_slides")
}
\arguments{
\item{indir}{Input dir, defaults to 'slides'}

\item{outdir}{Output dir, defaults to 'pdf_slides'}
}
\value{
Side effect; renders all the html to pdf
}
\description{
Render all html slide files in a folder (to depth 1) to pdf with decktape
}
\details{
By default will recurse to 1 level of subdirs
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
