% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{kjh_js_files}
\alias{kjh_js_files}
\title{Path to js file}
\usage{
kjh_js_files(js_file = "kjh-macros.js", subdir = "js")
}
\arguments{
\item{js_file}{Javascript file}

\item{subdir}{JS file directory}
}
\value{
Filepath to js file
}
\description{
Path to js file
}
