% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templating.R
\name{kjh_lib_dir}
\alias{kjh_lib_dir}
\title{The basic lib dir}
\usage{
kjh_lib_dir(libdir = "slides/libs")
}
\arguments{
\item{libdir}{Location of the libs directory}
}
\value{
path to libs
}
\description{
The basic lib dir
}
