% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_format.R
\name{kjh_local_slides_reader}
\alias{kjh_local_slides_reader}
\title{kjh Slide Document Format. Local version.}
\usage{
kjh_local_slides_reader(...)
}
\arguments{
\item{...}{Args to be passed on to xaringan::moon_reader}
}
\value{
Document format
}
\description{
A wrapper for \code{xaringan::moon_reader()} that sets various options and support file locations.
This version of the reader assumes all the files are local (i.e. in the top level of slides/)
}
