% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_slides.R
\name{kjh_purl_one_slide}
\alias{kjh_purl_one_slide}
\title{Purl one Rmd file}
\usage{
kjh_purl_one_slide(infile, outdir = "code")
}
\arguments{
\item{infile}{Absolute or relative Rmd file name}

\item{outdir}{Output directory, default 'code'}
}
\value{
purled file
}
\description{
Purl one Rmd file
}
