% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_slides.R
\name{kjh_purl_slides}
\alias{kjh_purl_slides}
\title{Purl all slide .Rmds to .R and put them in the \verb{code/} folder}
\usage{
kjh_purl_slides(indir = "slides", outdir = "code")
}
\arguments{
\item{indir}{The source dir, default slides, crawled recursively}

\item{outdir}{The output dir, default code}
}
\value{
Side effect; \verb{code/} folder of purled .R files.
}
\description{
Convert all Rmd files in the slide folder to R files in the code folder
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
