% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{kjh_register_myriad}
\alias{kjh_register_myriad}
\title{Register Myriad font variant}
\usage{
kjh_register_myriad()
}
\value{
Myriad variant registered in systemfonts database
}
\description{
Register Myriad font variant
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
