% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{kjh_register_tenso}
\alias{kjh_register_tenso}
\title{Register Tenso font variant}
\usage{
kjh_register_tenso()
}
\value{
Tenso variant registered in systemfonts databast
}
\description{
Register Tenso font variant
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
