% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{kjh_remark_js}
\alias{kjh_remark_js}
\title{Path to remark.js}
\usage{
kjh_remark_js(remark = "libs/remark-latest.min.js")
}
\arguments{
\item{remark}{Local path to remark in package inst files}
}
\value{
Absolute path to remark within slides
}
\description{
Path to remark.js
}
