% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{kjh_render_all_slides}
\alias{kjh_render_all_slides}
\title{Knit all files in the slides dir}
\usage{
kjh_render_all_slides(indir = "slides")
}
\arguments{
\item{indir}{Input directory (default 'slides')}
}
\value{
Side effect; Renders all the Rmd files to html
}
\description{
Knit all files in the slides dir
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
