% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_slides.R
\name{kjh_render_one_slide}
\alias{kjh_render_one_slide}
\title{Render one slide deck from Rmd to HTML}
\usage{
kjh_render_one_slide(infile, quietly = TRUE)
}
\arguments{
\item{infile}{Relative or absolute path to Rmd input file}

\item{quietly}{Run rmarkdown::render() quietly}
}
\value{
Rendered HTML file
}
\description{
Render one slide deck from Rmd to HTML
}
