% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{kjh_resource_file}
\alias{kjh_resource_file}
\title{Path to the kjhslides package}
\usage{
kjh_resource_file(file, subdir = NULL, libpath = "slides")
}
\arguments{
\item{file}{The file you want}

\item{libpath}{The path to slides support files}

\item{subpath}{Desired subfolder if any}
}
\value{
A file path
}
\description{
Path to the kjhslides package
}
