% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_theme.R
\name{kjh_set_classic_theme}
\alias{kjh_set_classic_theme}
\title{Set to classic theme}
\usage{
kjh_set_classic_theme(n = 5)
}
\arguments{
\item{n}{N colors for color theme}
}
\value{
Sets the theme to the ggplot theme_classic
}
\description{
Set to classic theme
}
