% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{kjh_set_showtext}
\alias{kjh_set_showtext}
\title{Turn on showtext}
\usage{
kjh_set_showtext(dpi = 300)
}
\arguments{
\item{dpi}{DPI for output (300 default is recommended)}
}
\value{
Turns on showtext
}
\description{
Turn on showtext
}
