% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{kjh_set_slide_theme}
\alias{kjh_set_slide_theme}
\title{Slide theme}
\usage{
kjh_set_slide_theme()
}
\value{
Environment vars set and theme set etc
}
\description{
Set up slides theme
}
\details{
Sets up the slide theme with Tenso font, Okabe-Ito colors, and element_markdown()
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
