% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{kjh_set_xaringan_opts}
\alias{kjh_set_xaringan_opts}
\title{Set Xaringan Options}
\usage{
kjh_set_xaringan_opts()
}
\value{
Slide setup
}
\description{
Set Xaringan Options
}
\details{
Sets Tile view, animate css, fade transition, use clipboard.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
