% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_theme.R
\name{kjh_theme_map}
\alias{kjh_theme_map}
\title{Provide theme map}
\usage{
kjh_theme_map()
}
\value{
them_map
}
\description{
Provide theme map
}
\details{
Working around flippbookr's issues with finding things
}
