% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_theme.R
\name{kjh_theme_nymap}
\alias{kjh_theme_nymap}
\title{NYC Map theme}
\usage{
kjh_theme_nymap()
}
\value{
NYC map theme
}
\description{
NYC Map theme
}
\details{
Work around flipbookr's trouble finding things
}
