% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_processing.R
\name{kjh_unname_chunks}
\alias{kjh_unname_chunks}
\title{Unname all unnameable chunks}
\usage{
kjh_unname_chunks(path, chunk_name_prefix = NULL)
}
\arguments{
\item{path}{Path to Rmd or qmd file}

\item{chunk_name_prefix}{Character string with prefix of chunknames that will be removed.
Default: NULL (indicating all chunknames will be removed except the one named 'setup' and
any chunks with the include = FALSE option, to protect flipbookr references.)}
}
\value{
Unnames chunks, except setup and include=FALSE chunks
}
\description{
Unname chunks except 'setup' and flipbookr chunks that have include=FALSE
}
