% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{ragg_png}
\alias{ragg_png}
\title{Set graphics device to ragg PNG}
\usage{
ragg_png(..., res = 150)
}
\arguments{
\item{...}{Passed on to \code{agg_png()}}

\item{res}{PNG resolution, defaults to 150}
}
\value{
Sets the device
}
\description{
Set graphics device to ragg PNG
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
