% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{setup_slides}
\alias{setup_slides}
\title{setup_slides}
\usage{
setup_slides(
  folder,
  zipfile = "slides.zip",
  slidefolder = "slides",
  destfolder = "slides"
)
}
\arguments{
\item{folder}{The destination to copy to within the user's home.
This must be supplied by the user.}

\item{zipfile}{The name of the bundled slide file.}

\item{slidefolder}{The name of the course packet folder to be created}
}
\value{
The \code{zipfile} is copied to \code{folder} and its contents
expanded into a directory, the \code{slidefolder}.
}
\description{
Copy slide infrastructure to a given folder
}
\details{
Transfers a zip file containing course materials from the socviz
library to the Desktop.
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
setup_slides()
}
\author{
Kieran Healy
}
