% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slide_theme.R
\name{theme_tenso}
\alias{theme_tenso}
\title{Theme Tenso}
\usage{
theme_tenso(base_size = 12, base_family = "Tenso Slide")
}
\arguments{
\item{base_size}{Font size}

\item{base_family}{Font Family (Tenso Slide)}
}
\value{
Slide ggplot Theme
}
\description{
Theme Tenso
}
