% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{kjh_decktape_one}
\alias{kjh_decktape_one}
\title{Convert an HTML slide deck to PDF}
\usage{
kjh_decktape_one(infile, outdir = "pdf_slides")
}
\arguments{
\item{infile}{Input html file}

\item{outdir}{Output directory, defaults to \verb{pdf_slides/}}
}
\value{
Side-effect; rendered PDF slide file
}
\description{
Render one slide html file to pdf_slides/file.pdf with decktape
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
