% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_format.R
\name{kjh_slides_reader}
\alias{kjh_slides_reader}
\title{kjh Slide Document Format.}
\usage{
kjh_slides_reader(...)
}
\arguments{
\item{...}{Args to be passed on to xaringan::moon_reader}
}
\value{
Document format
}
\description{
A wrapper for \code{xaringan::moon_reader()} that sets various options and support file locations.
}
