% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunk_processing.R
\name{kjh_unname_all_chunks}
\alias{kjh_unname_all_chunks}
\title{Unname all chunks}
\usage{
kjh_unname_all_chunks(indir = "slides")
}
\arguments{
\item{indir}{The input directory, default "slides"}
}
\value{
All the Rmd files with their chunk names removed
}
\description{
Unname all Rmd chunks in the Rmd folders
}
\details{
Recurses 1 level (i.e. subdirs) by default. Right now this will break any references made with chunk_reveal!
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
