% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.r
\name{kjh_update_all_chunk_names}
\alias{kjh_update_all_chunk_names}
\title{Convert all chunks}
\usage{
kjh_update_all_chunk_names(indir = "slides")
}
\arguments{
\item{indir}{Input directory, default "slides"}
}
\value{
All Rmd chunks are unnamed and then renamed
}
\description{
Run unname and then rename on all chunks in all files
}
\details{
Recurses to 1 level of depth (i.e. subdirs).
Right now this will break any references made with chunk_reveal!
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
